+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/nodeInfo/PeerType.java
@@ -1,0 +1,16 @@
+package com.concordium.sdk.responses.nodeInfo;
+
+/**
+ * Represents type of Node.
+ */
+public enum PeerType {
+    /**
+     * Node of type Bootstrapper.
+     */
+    Bootstrapper,
+
+    /**
+     * Node of type Node.
+     */
+    Node,
+}